rpi/requirements.txt
flask~=3.1.1
sqlalchemy~=2.0.41
apscheduler~=3.11.0
pyserial~=3.5
colorama~=0.4.6
Flask-Migrate~=4.1.0
alembic~=1.16.4